add a generic call-tokenuri strategy
The purpose of this generic call-tokenuri strategy is to remove the
need of diffferent strategies such as soundxyz-call-tokenuri,
zora-call-tokenuri, catalog-call-tokenuri etc.

The benefits include:
- Less number of files to mantain
- Easier to refactor
- Easier to integrate new indie contracts

This new call-tokenuri strategy has been modified to not need
logs-to-subgraph strategy.

I have tested this commit and the output files are the same as before.